Move success to own package
