+++ src/entries/components/flows/common/SuccessNotice/SuccessNotice.module.scss
