+++ src/entries/components/flows/common/SuccessNotice/SuccessNotice.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Success.module.scss';
+import styles from './SuccessNotice.module.scss';
 import successImage from './success.svg';
 
 export const SuccessNotice: React.FunctionComponent<{ children?: React.ReactNode }> = ({
